requirements.txt
Django==5.2
pillow==11.2.1
requests==2.32.3
python-dotenv==1.1.0
django-admin-sortable2==2.2.6
django-tinymce==4.1.0
